move: API 응답을 위한 *Result 클래스 패키지 이동
