+++ src/main/java/com/programmers/board/controller/PostController.java
@@ -1,5 +1,7 @@
 package com.programmers.board.controller;
 
+import com.programmers.board.controller.response.PageResult;
+import com.programmers.board.controller.response.Result;
 import com.programmers.board.dto.PostDto;
 import com.programmers.board.dto.request.PostCreateRequest;
 import com.programmers.board.dto.request.PostUpdateRequest;
